live: move ISO embed area infrastructure to a separate module
